Correccion de error al eliminar

+++ src/main/java/com/co/quality/clothing/services/impl/ProductosServiceImpl.java
@@ -117,9 +117,12 @@ public class ProductosServiceImpl implements ProductosService {
 
             repository.deleteById(id);
 
-            for (FileMetadata imagen : producto.getImagenes()) {
-                fileMetadataService.deleteArchivo(imagen.getId());
+            if(producto.getImagenes() != null) {
+                for (FileMetadata imagen : producto.getImagenes()) {
+                    fileMetadataService.deleteArchivo(imagen.getId());
+                }
             }
+
             return ResponseEntity.noContent().build();
         }
         return ResponseEntity.notFound().build();
